python/requirements.txt
pyserial
matplotlib
pandas
numpy
